+++ Changelog.md
@@ -8,6 +8,7 @@
 - move licenses to own credits page
 - move devmode to credits page
 - move import/export to new backup page
+- improve backup page
 
 ## 0.4.0