+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.5.0
+
 - add Stats for time on days in week/month
 - Introducing the Statistics tab
 - make "simons bot format" an devmode option (hidden by default)